+++ include/boost/algorithm/clamp.hpp
@@ -27,7 +27,7 @@
 #include <boost/range/begin.hpp>
 #include <boost/range/end.hpp>
 #include <boost/type_traits/type_identity.hpp> // for boost::type_identity
-#include <boost/utility/enable_if.hpp>         // for boost::disable_if
+#include <boost/core/enable_if.hpp>            // for boost::disable_if
 
 namespace boost { namespace algorithm {
