+++ include/boost/algorithm/cxx11/is_sorted.hpp
@@ -20,7 +20,7 @@
 #include <boost/range/begin.hpp>
 #include <boost/range/end.hpp>
 
-#include <boost/utility/enable_if.hpp>
+#include <boost/core/enable_if.hpp>
 #include <boost/type_traits/is_same.hpp>
 #include <boost/type_traits/type_identity.hpp> // for boost::type_identity
